+++ public/sw.js
@@ -1,4 +1,4 @@
-const CACHE_NAME = 'cok-static-v1'
+const CACHE_NAME = 'cok-static-v2'
 const APP_SHELL = ['./', './index.html', './manifest.webmanifest', './favicon.svg']
 
 self.addEventListener('install', (event) => {